Move Wiki class out of init file
